build(Wrappers) Remove wrappers coverage files
Criterion creates a shit load of branches I can't test. And wrap_malloc's are more
of an example to people who want to use it anyway, so make tests now removes
wrappers' coverage files in order to not get reduced branch coverage on reports
for stuff I literally can't test.